making bash use arguments

+++ bash/git-push.sh
@@ -1,14 +1,16 @@
 #!/run/current-system/sw/bin/bash
 
-rm -f ~/.modular-nixos-configuration/users/user.nix 
-
-git add -A
-
-read -p 'Message: ' msg
-git commit -m "$msg"
-
-read -p 'Repository: ' repo
-read -p 'Branch: ' branch
-git push -u $repo $branch
-
-~/.modular-nixos-configuration/bash/userQuery.sh
+if [ "$1" = "--help" ]; then
+    echo "git-push.sh <user> <msg> <repo> <branch>"
+fi
+if [ "$1" = "timo" ] || [ "$1" = "matteo" ]; then 
+    rm -f ~/.modular-nixos-configuration/users/user.nix 
+    
+    git add -A
+    
+    git commit -m "$2"
+    
+    git push -u $3 $4
+    
+    ~/.modular-nixos-configuration/bash/userQuery.sh $1
+fi
